requirements.txt
astroid==2.4.2
Click==7.0
dnspython==1.16.0
eventlet==0.25.1
Flask==1.1.1
Flask-SocketIO==4.2.1
greenlet==0.4.15
gunicorn==19.9.0
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
monotonic==1.5
pylint==2.6.0
python-dotenv==0.15.0
python-engineio==3.9.3
python-socketio==4.3.1
six==1.12.0
toml==0.10.1
Werkzeug==0.16.0
wrapt==1.12.1
